Simplify code by using the POINTS constant in the iterator
We could directly iterate and use the POINTS array because if guesses
or winners' count is more than the POINTS array length it is useless
to check them. However due to the fact that the winners and/or guesses
could have less length than the POINTS array an additional check is added.